ArcGIS duplicate features sent to the feature service (#259)
* Fixes issue where deleted/archived observations were consistently 'deleted' from the feature layer. Other code clean up.

* Updated github actions to build docker container on feature branches. Will still only deploy to dev/test on develop/test branch.

---------
