+++ .github/workflows/aws_dev.deploy.yaml
@@ -3,8 +3,8 @@ name: AWS Dev Deployment
 on:
   push:
     branches:
-      - develop
-      - test
+      - '**'
+  workflow_dispatch:
 
 permissions:
   id-token: write
@@ -36,17 +36,18 @@ jobs:
         id: ecr-login
         uses: aws-actions/amazon-ecr-login@v2
 
-      - name: Build Docker image and push
+      - name: Build Docker image
         env:
           REGISTRY: ${{ steps.ecr-login.outputs.registry }}
           REPOSITORY: "/mage/server"
           IMAGE_TAG: ${{ env.IMAGE_TAG }}
         run: |
           docker build -t $REGISTRY$REPOSITORY:$IMAGE_TAG .
-          docker push $REGISTRY$REPOSITORY:$IMAGE_TAG
 
-      - name: ReDeploy ECS Service
+      - name: Push Image and ReDeploy ECS Service
+        if: github.ref =='refs/heads/test' || github.ref =='refs/heads/develop'
         env:
           SERVICE: "mage${{ env.IMAGE_TAG }}"
         run: |
+          docker push $REGISTRY$REPOSITORY:$IMAGE_TAG
           aws ecs update-service --cluster magedev --service $SERVICE --force-new-deployment
