+++ plugins/arcgis/service/src/FeatureLayerProcessor.ts
@@ -2,7 +2,6 @@ import { ArcGISPluginConfig } from "./types/ArcGISPluginConfig";
 import { ArcObjects } from "./ArcObjects";
 import { FeatureQuerier } from "./FeatureQuerier";
 import { LayerInfo } from "./LayerInfo";
-import { ObservationBinner } from "./ObservationBinner";
 import { ObservationBins } from "./ObservationBins";
 import { ObservationsSender } from "./ObservationsSender";
 import { ArcGISIdentityManager } from "@esri/arcgis-rest-request";
@@ -17,11 +16,6 @@ export class FeatureLayerProcessor {
     layerInfo: LayerInfo;
 
     /**
-     * Bins the observations into updates and adds.
-     */
-    private _binner: ObservationBinner;
-
-    /**
      * Sends the observation adds or updates to the arc feature layer.
      */
     sender: ObservationsSender;
@@ -34,7 +28,26 @@ export class FeatureLayerProcessor {
     /**
      * The last time we checked for new/modified observations.
      */
-    lastTimeStamp: number;
+    lastTimeStamp: number = 0;
+
+    /**
+     * The number of existence queries we are still waiting for.
+     */
+    private _existenceQueryCounts: number = 0;
+
+    /**
+     * Contains the results from checking if an observation exists on the server.
+     */
+    private _pendingNewAndUpdates: ObservationBins;
+
+    /**
+     * The plugins configuration.
+     */
+    private _config: ArcGISPluginConfig;
+
+    private _console: Console;
+
+    private _addedObs: Set<string> = new Set<string>();
 
     /**
      * Creates a new instance of FeatureLayerProcessor.
@@ -45,26 +58,38 @@ export class FeatureLayerProcessor {
      */
     constructor(layerInfo: LayerInfo, config: ArcGISPluginConfig, identityManager: ArcGISIdentityManager, console: Console) {
         this.layerInfo = layerInfo;
-        this.lastTimeStamp = 0;
+        this._config = config;
+        this._console = console;
         this.featureQuerier = new FeatureQuerier(layerInfo, config, identityManager, console);
-        this._binner = new ObservationBinner(layerInfo, this.featureQuerier, config);
         this.sender = new ObservationsSender(layerInfo, config, identityManager, console);
+        this._pendingNewAndUpdates = new ObservationBins;
     }
 
     /**
-     * Indicates if this processor has pending updates still waiting to be processed.
+     * Indicates if this binner has pending updates still waiting to be processed.
      * @returns {boolean} True if it is still waiting for updates to be processed, false otherwise.
      */
     hasPendingUpdates(): boolean {
-        return this._binner.hasPendingUpdates();
+        return this._existenceQueryCounts > 0;
     }
 
     /**
      * Checks to see if there are any updates that need to be sent to the feature layer.
      */
     async processPendingUpdates() {
-        const bins = this._binner.pendingUpdates();
-        await this.send(bins);
+        const newAndUpdates = new ObservationBins();
+        for (let i = 0; i < this._pendingNewAndUpdates.adds.count(); i++) {
+            if (!this._addedObs.has(this._pendingNewAndUpdates.adds.observations[i].id)) {
+                newAndUpdates.adds.add(this._pendingNewAndUpdates.adds.observations[i]);
+                this._addedObs.add(this._pendingNewAndUpdates.adds.observations[i].id);
+            }
+        }
+        for (let i = 0; i < this._pendingNewAndUpdates.updates.count(); i++) {
+            newAndUpdates.updates.add(this._pendingNewAndUpdates.updates.observations[i]);
+        }
+        this._pendingNewAndUpdates.clear();
+
+        await this.send(newAndUpdates);
     }
 
     /**
@@ -81,12 +106,65 @@ export class FeatureLayerProcessor {
             }
         }
 
-        const bins = this._binner.sortEmOut(arcObjectsForLayer);
+        const bins = new ObservationBins();
+
+        for (const arcObservation of arcObjectsForLayer.observations) {
+            // TODO: Would probably want a better way to determine which observations need to be updated in arcgis
+            if (arcObjectsForLayer.firstRun || arcObservation.lastModified !== arcObservation.createdAt) {
+                bins.updates.add(arcObservation);
+            } else if (!this._addedObs.has(arcObservation.id)) {
+                bins.adds.add(arcObservation);
+                this._addedObs.add(arcObservation.id);
+            }
+        }
+
+        for (const arcObservation of bins.updates.observations) {
+            this._existenceQueryCounts++;
+            this.featureQuerier.queryObservation(arcObservation.id, (result) => {
+                this._existenceQueryCounts--;
+                if (result.features != null && result.features.length > 0) {
+                    this._addedObs.add(arcObservation.id);
+                    const arcAttributes = result.features[0].attributes;
+
+                    let lastEdited = null;
+                    if (this._config.lastEditedDateField != null) {
+                        lastEdited = Number(arcAttributes[this._config.lastEditedDateField]);
+                    }
+                    if (!lastEdited || lastEdited < arcObservation.lastModified) {
+
+                        const objectIdField = result.objectIdFieldName;
+                        const updateAttributes = arcObservation.object.attributes;
+
+                        updateAttributes[objectIdField] = arcAttributes[objectIdField];
+
+                        // Determine if any editable attribute values should be deleted
+                        const lowerCaseUpdateAttributes = Object.fromEntries(
+                            Object.entries(updateAttributes).map(([k, v]) => [k.toLowerCase(), v])
+                        );
+                        for (const arcAttribute of Object.keys(arcAttributes)) {
+                            if (this.layerInfo.isEditable(arcAttribute)
+                                && arcAttributes[arcAttribute] != null
+                                && lowerCaseUpdateAttributes[arcAttribute.toLowerCase()] == null) {
+                                updateAttributes[arcAttribute] = null;
+                            }
+                        }
+
+                        this._pendingNewAndUpdates.updates.add(arcObservation);
+                    }
+
+                } else {
+                    this._pendingNewAndUpdates.adds.add(arcObservation);
+                }
+            }, undefined, false);
+        }
+        bins.updates = this._pendingNewAndUpdates.updates;
+
         await this.send(bins);
 
         for (const arcObservation of observations.deletions) {
-            if (this.layerInfo.geometryType === arcObservation.esriGeometryType) {
+            if (this.layerInfo.geometryType === arcObservation.esriGeometryType && this._addedObs.has(arcObservation.id)) {
                 await this.sender.sendDelete(arcObservation.id);
+                this._addedObs.delete(arcObservation.id);
             }
         }
     }
